+++ src/black/linegen.py
@@ -561,7 +561,7 @@ def transform_line(
     if (
         not line.contains_uncollapsable_type_comments()
         and not line.should_split_rhs
-        and not line.magic_trailing_comma
+        and not line.bracket_after_magic_trailing_comma
         and (
             is_line_short_enough(line, mode=mode, line_str=line_str)
             or line.contains_unsplittable_type_ignore()
@@ -619,6 +619,24 @@ def transform_line(
                     string_paren_wrap,
                     rhs,
                 ]
+
+                # If there are comments at the beginning or end of the line, call
+                # standalone_comment_split before delimiter_split. Else, call
+                # delimiter_split first. This means that when there is a comment in the
+                # middle of an expression or list, it will perform a delimiter split.
+                # But when the comment is at the top or bottom, it won't perform a
+                # delimiter split (if it doesn't have to).
+                if (
+                    Preview.stop_some_unnecessary_wrapping_inside_brackets in mode
+                    and line.contains_standalone_comments()
+                    and len(line.comments) == 0
+                    and all(
+                        leaf.type != STANDALONE_COMMENT for leaf in line.leaves[1:-1]
+                    )
+                    and line.magic_trailing_comma is None
+                ):
+                    transformers.remove(standalone_comment_split)
+                    transformers.insert(0, standalone_comment_split)
             else:
                 transformers = [
                     string_merge,
@@ -800,6 +818,7 @@ def _first_right_hand_split(
     body = bracket_split_build_line(
         body_leaves, line, opening_bracket, component=_BracketSplitComponent.body
     )
+    body.magic_trailing_comma = line.magic_trailing_comma
     tail = bracket_split_build_line(
         tail_leaves, line, opening_bracket, component=_BracketSplitComponent.tail
     )
@@ -848,7 +867,7 @@ def _maybe_split_omitting_optional_parens(
                 )
                 # the left side of assignment won't explode further because of magic
                 # trailing comma
-                and rhs.head.magic_trailing_comma is None
+                and rhs.head.bracket_after_magic_trailing_comma is None
                 # the split by omitting optional parens isn't preferred by some other
                 # reason
                 and not _prefer_split_rhs_oop(rhs_oop, mode)
@@ -1526,7 +1545,7 @@ def generate_trailers_to_omit(line: Line, line_length: int) -> Iterator[Set[Leaf
     """
 
     omit: Set[LeafID] = set()
-    if not line.magic_trailing_comma:
+    if not line.bracket_after_magic_trailing_comma:
         yield omit
 
     length = 4 * line.depth