+++ src/black/mode.py
@@ -185,6 +185,7 @@ class Preview(Enum):
     skip_magic_trailing_comma_in_subscript = auto()
     wrap_long_dict_values_in_parens = auto()
     wrap_multiple_context_managers_in_parens = auto()
+    stop_some_unnecessary_wrapping_inside_brackets = auto()
     dummy_implementations = auto()
     walrus_subscript = auto()