+++ tests/data/miscellaneous/skip_magic_trailing_comma.py
@@ -1,0 +1,5 @@
+def func(
+    x,
+    # this comment shouldn't be deleted
+):
+    pass